requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-ollama==0.3.3
langchain-experimental==0.3.4
neo4j==5.28.1
tiktoken==0.9.0
python-dotenv==1.1.0
json-repair==0.46.2
langchain-openai==0.3.21
langchain_core==0.3.64
numpy==2.2.6
pydantic==2.11.5